+++ xnt-core/src/protocol/consensus/transaction/transaction_proof.rs
@@ -155,7 +155,8 @@ impl TransactionProof {
                 ConsensusRuleSet::TimelockExtension
                 | ConsensusRuleSet::UpgradeVM
                 | ConsensusRuleSet::UpgradeVMv4
-                | ConsensusRuleSet::UpgradeVMv5 => {
+                | ConsensusRuleSet::UpgradeVMv5
+                | ConsensusRuleSet::UpgradeVMv7 => {
                     proof_collection.verify_v2(kernel_mast_hash, network).await
                 }
             },